requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
google-genai>=1.0.0
openai>=1.0.0
rich>=13.0.0
tabulate>=0.9.0
gspread>=6.0.0
google-auth>=2.20.0
flask>=3.0.0

# MCP client (for Step 1/2/3 architecture)
mcp>=1.8.0,<2.0.0
